backend/requirements.txt
alembic==1.0.10
appdirs==1.4.3
attrs==19.1.0
bcrypt==3.1.6
black==19.3b0
cffi==1.12.3
Click==7.0
coverage==4.5.3
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
nose2==0.9.1
python-dotenv==0.10.2
six==1.12.0
SQLAlchemy==1.3.3
toml==0.10.0
Werkzeug==0.15.4
WTForms==2.2.1
web3
